get the damn thing working again!
